Plugin MVP
I'm considering adding custom block with advanced functionality. Best practices usually suggests keeping this code in a plugin rather than a theme, so the functionality remains after switching to another theme. I don't think that's a likely scenario for this project, but it'd be a good exercise anyway. This is the MVP for getting a plugin to show up in WP's Plugins page and activate, and show a new item in the block editor's block selection dialogue.