Block Theme (Cloned from Twenty Twenty-Four) MVP
Big paradigm shift: I started out with the intention of creating my own block theme from scratch, but with the pace in which WordPress's own spec is evolving, that is proving to be prohibitive. So, I used the CreateBlockTheme plugin to clone the Twenty Twenty-Four theme so I'd have a solid starting-point, and that's been going well so far, so here's a commit.