Add Function: Open Links in New Tab
Some of WordPress' blocks allow you to specify that a link should open in a new tab or window, but not all. This custom function should automatically apply the "open in new tab/window" behavior to any link inside an element that is given the CSS class "open-link-in-new-tab" (or whatever string is specified inside the function.